requirements.txt
fastapi
uvicorn[standard]
pymongo
motor
langchain
langchain-google-genai
firebase-admin
python-dotenv
pydantic
PyPDF2
tiktoken
google-generativeai